// bmstu-hospital-client/src/api/models/Speciality.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Snippets API
 * Test description
 *
 * The version of the OpenAPI document: v1
 * Contact: contact@snippets.local
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Speciality
 */
export interface Speciality {
    /**
     * 
     * @type {number}
     * @memberof Speciality
     */
    readonly id?: number;
    /**
     * 
     * @type {string}
     * @memberof Speciality
     */
    readonly url?: string;
    /**
     * 
     * @type {string}
     * @memberof Speciality
     */
    name: string;
    /**
     * 
     * @type {Set<number>}
     * @memberof Speciality
     */
    readonly doctors?: Set<number>;
    /**
     * 
     * @type {string}
     * @memberof Speciality
     */
    description: string;
}

/**
 * Check if a given object implements the Speciality interface.
 */
export function instanceOfSpeciality(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "description" in value;

    return isInstance;
}

export function SpecialityFromJSON(json: any): Speciality {
    return SpecialityFromJSONTyped(json, false);
}

export function SpecialityFromJSONTyped(json: any, ignoreDiscriminator: boolean): Speciality {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'url': !exists(json, 'url') ? undefined : json['url'],
        'name': json['name'],
        'doctors': !exists(json, 'doctors') ? undefined : json['doctors'],
        'description': json['description'],
    };
}

export function SpecialityToJSON(value?: Speciality | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'description': value.description,
    };
}